src/main/java/com/model/Book.java
package com.model;

import java.util.Objects;


public class Book {
private String bookID;
private String title;
private String author;
private String category;
private double price;
public Book(String bookID,String title, String author, String category,double price){
	this.bookID=bookID;
	this.title=title;
	this.author=author;
	this.category=category;
	this.price=price;
}
public String getBookID() {
	return bookID;
}
public void setBookID(String bookID) {
	this.bookID = bookID;
}
public String getTitle() {
	return title;
}
public void setTitle(String title) {
	this.title = title;
}
public String getAuthor() {
	return author;
}
public void setAuthor(String author) {
	this.author = author;
}
public String getCategory() {
	return category;
}
public void setCategory(String category) {
	this.category = category;
}
public double getPrice() {
	return price;
}
public void setPrice(double price) {
	this.price = price;
}
@Override 
public String toString() {
	return "BookId-"+getBookID()+" Book name-"+getTitle()+" Author-"+getAuthor()+" Category-"+getCategory()+" Price-"+getPrice();
}
}
